improve code eligibility
