+++ etl/dca_operatingbusinesses.py
@@ -29,6 +29,14 @@ dca_operatingbusinesses = Flow(
     filter_rows(not_equals = [
             dict(address_borough = 'Outside NYC'),
             ]),
+    
+    filter_rows(equals = [
+            dict(industry = 'Scrap Metal Processor'),
+            dict(industry = 'Parking Lot'),
+            dict(industry = 'Garage'),
+            dict(industry = 'Garage and Parking Lot'),
+            dict(industry = 'Tow Truck Company'),
+            ]),
             
     add_field('datasource', 'string', table_name),
 
@@ -56,14 +64,9 @@ dca_operatingbusinesses = Flow(
                             ]),
 
     geo_flow,
-    add_computed_field([dict(target=dict(name = 'the_geom_tmp', type = 'string'),
+    add_computed_field([dict(target=dict(name = 'the_geom', type = 'string'),
                             operation=lambda row: get_the_geom(row['geo_longitude'], row['geo_latitude'])
                             ),
-                        dict(target=dict(name = 'the_geom', type = 'string'),
-                            operation=lambda row: row['the_geom_tmp']
-                                if row['the_geom_tmp'] != None
-                                else get_the_geom(row['longitude'], row['latitude']) 
-                            )
                         ]),
     dump_to_postgis(table_name)
 )
