+++ etl/dcas_colp.py
@@ -12,25 +12,9 @@ csv.field_size_limit(sys.maxsize)
 table_name = 'dcas_colp'
 dcas_colp = Flow(
     load(url, resources = table_name, force_strings=False),
-    # cacheing table
+
     checkpoint(table_name),
-    # uid
-    # facname
-    # factype
-    # facsubgrp
-    # facgroup
-    # facdomain
-    # servarea
-    # opname
-    # opabbrev
-    # optype
-    # overagency
-    # overabbrev
-    # overlevel
-    # capacity
-    # captype
-    # proptype
-    # datasource
+
     add_field('datasource', 'string', table_name),
 
     ################## geospatial ###################
@@ -38,20 +22,17 @@ dcas_colp = Flow(
     ###### exist before geo_flows          ########## 
     #################################################
 
-    # hnum
-    rename_field('house_number', 'hnum'),
-    # sname
-    rename_field('street_name', 'sname'), 
-    # zipcode
     add_field('zipcode', 'string', ''),
-    # boro
+    rename_field('house_number', 'hnum'),
+    rename_field('street_name', 'sname'), 
     rename_field('borough', 'boro'), 
+
     geo_flow,
     add_computed_field([dict(target=dict(name = 'the_geom', type = 'string'),
                             operation=lambda row: get_the_geom(row['geo_longitude'], row['geo_latitude'])
                             )
                         ]),
-    printer(num_rows=3),
+                        
     dump_to_postgis(),
 )
