+++ etl/dfta_contracts.py
@@ -12,19 +12,15 @@ csv.field_size_limit(sys.maxsize)
 table_name = 'dfta_contracts'
 dfta_contracts = Flow(
     load(url, resources = table_name, force_strings=False),
-    # cacheing table
     checkpoint(table_name),
-    # datasource
-    add_field('datasource', 'string', table_name),
 
+    add_field('datasource', 'string', table_name),
 
     ################## geospatial ###################
     ###### Make sure the following columns ##########
     ###### exist before geo_flows          ########## 
     #################################################
 
-    # hnum
-    # sname 
     add_computed_field([dict(target=dict(name = 'address', type = 'string'),
                                         operation=lambda row: quick_clean(row['program_address'])
                                         ),
@@ -35,9 +31,9 @@ dfta_contracts = Flow(
                                     operation=lambda row: get_sname(row['address'])
                                     )
                         ]),
-    # boro
+
     add_field('boro', 'string', ''),
-    # zipcode
+
     rename_field('program_zipcode', 'zipcode'),
 
     geo_flow,
@@ -45,7 +41,7 @@ dfta_contracts = Flow(
                             operation=lambda row: get_the_geom(row['geo_longitude'], row['geo_latitude'])
                             )
                         ]),
-    printer(num_rows=3),
+    # printer(num_rows=3),
     dump_to_postgis()
 )
