+++ etl/utils.py
@@ -24,24 +24,13 @@ fields = ['uid', 'facname',
         'council', 'censtract', 
         'datasource', 'geom']
 
-url = 'https://db-data-recipes.sfo2.digitaloceanspaces.com/pipelines/db-facilities/2019-05-17/datapackage.json'
+url = 'https://db-data-recipes.sfo2.digitaloceanspaces.com/pipelines/db-facilities/2019-05-21/datapackage.json'
 
 def get_the_geom(lon, lat): 
         lon = float(lon) if lon != '' else None
         lat = float(lat) if lat != '' else None
         if (lon is not None) and (lat is not None): 
                 return str(Point(lon, lat))
-
-def get_geom_source(s):
-        try:
-                s = s[(s.find('(')):]
-                lat = float(s[1:s.find(',')])
-                lon = float(s[s.find(',')+1:-1])
-                if (lon is not None) and (lat is not None): 
-                        return str(Point(lon, lat))
-        except:
-                return ''
-
 
 def quick_clean(address):
         address = '-'.join([i.strip() for i in address.split('-')])
